feat: add task list UI with search, filter chips, and swipe-to-delete

+++ app/src/main/java/com/example/collegecompanion/data/local/TaskDao.kt
@@ -1,3 +1,4 @@
+// data/local/dao/TaskDao.kt
 package com.example.collegecompanion.data.local
 
 import androidx.room.*
@@ -7,17 +8,16 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface TaskDao {
 
-    // ── Read ──────────────────────────────────────────────────────────────
-    @Query("SELECT * FROM tasks ORDER BY id DESC")          // removed createdAt
+    @Query("SELECT * FROM tasks ORDER BY id DESC")
     fun getAllTasks(): Flow<List<Task>>
 
-    @Query("SELECT * FROM tasks WHERE id = :id LIMIT 1")    // removed duplicate
-    suspend fun getTaskById(id: Long): Task?                 // Long, not Int
+    @Query("SELECT * FROM tasks WHERE id = :id LIMIT 1")
+    suspend fun getTaskById(id: Int): Task?
 
     @Query("SELECT * FROM tasks WHERE isCompleted = 0 ORDER BY dueDate ASC")
     fun getPendingTasks(): Flow<List<Task>>
 
-    @Query("SELECT * FROM tasks WHERE isCompleted = 1 ORDER BY id DESC")  // removed createdAt
+    @Query("SELECT * FROM tasks WHERE isCompleted = 1 ORDER BY id DESC")
     fun getCompletedTasks(): Flow<List<Task>>
 
     @Query("SELECT * FROM tasks WHERE title LIKE :query")
@@ -26,7 +26,6 @@ interface TaskDao {
     @Query("SELECT * FROM tasks WHERE dueDate BETWEEN :startOfDay AND :endOfDay")
     fun getTasksDueToday(startOfDay: Long, endOfDay: Long): Flow<List<Task>>
 
-    // ── Write ─────────────────────────────────────────────────────────────
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertTask(task: Task): Long
 
@@ -37,7 +36,7 @@ interface TaskDao {
     suspend fun deleteTask(task: Task)
 
     @Query("UPDATE tasks SET isCompleted = :isCompleted WHERE id = :taskId")
-    suspend fun updateTaskCompletion(taskId: Long, isCompleted: Boolean) // Long, not Int
+    suspend fun updateTaskCompletion(taskId: Int, isCompleted: Boolean)
 
     @Query("DELETE FROM tasks WHERE isCompleted = 1")
     suspend fun deleteAllCompletedTasks()
